pyproject.toml
[tool.poetry]
name = "speakease-api"
version = "0.1.0"
description = ""
authors = ["gubo2012 <gubo2012v@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
uvicorn = "^0.34.3"
firebase-admin = "^6.9.0"
google-cloud-secret-manager = "^2.24.0"
google-generativeai = "^0.8.5"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
